Pdfassistant/requirements.txt
phidata
python-dotenv
yfinance
packaging
duckduckgo-search
fastapi
uvicorn
groq
openai
newspaper3k
sqlalchemy
pgvector
psycopg[binary]
pypdf
typer